Use derive macro for table rows

+++ src/index.rs
@@ -3,7 +3,7 @@ use std::hash::Hash;
 
 /// Index identifiers from a data type
 pub struct IdIndex<K> {
-  map: HashMap<K,u64>
+  map: HashMap<K,u32>
 }
 
 impl <K> IdIndex<K> where K: Eq + Hash {
@@ -20,8 +20,8 @@ impl <K> IdIndex<K> where K: Eq + Hash {
   }
 
   /// Get the ID for a key, adding it to the index if needed.
-  pub fn intern(&mut self, key: K) -> u64 {
-    let n = self.map.len() as u64;
+  pub fn intern(&mut self, key: K) -> u32 {
+    let n = self.map.len() as u32;
     *self.map.entry(key).or_insert(n + 1)
   }
 }
